Adjust top camera bar size
An alternative to `SafeAreaView` is proposed as status bar does not fit on IPhone 7 but
somewhat fit on IPhone X. So it is decided that status bar should be invisible.

However, `expo` could result in a peculiar bug where the constant
of the status bar height would become zero when the status bar is turned off
and the App is reloaded. Some hacky work-around is proposed to fight against
this weird bug.

It is estimated that such queer bug should not live in the production enviroment.
But for sake of development, a hacky component called `HikariStatusBar` gives its
birth.